requirements.txt
# Core dependencies for running automl_lib (CLI/pipeline/phases).
pydantic>=2.0
PyYAML>=6.0
numpy>=1.23
pandas>=1.5
scikit-learn>=1.3
matplotlib>=3.7
joblib>=1.3

# Config composition / structured configs (recommended).
hydra-core>=1.3
omegaconf>=2.3

# ClearML integration (optional but recommended for this project).
clearml>=1.14

# Optional visualization (used when available; code falls back if missing).
plotly>=5.0
